Reformat files with clang-format
Had to disable formatting around the Windows DLL export checks because
it would mess with the indentations on the function signatures after.
Hoping to find a better resolution to this issue, so that we dont need
to have hacky disable checks in place. We may even need to move our code
around.