Fix fill_boxes in referencing
This should have been a vec2f[], not a pointer. This forced the function
to only fill the first spot in the array, forgetting the rest.

It turns out the stream last week led me in the wrong direction for the
sake of empty filling for TDD. Thanks to subpixel in the Discord for
spotting my mistake.