SPRPD-5: Display level summaries before run
This allows players to see the level requirements before starting
the run to try and optimize their play. This UI required refactoring
the game's codebase to handle multiple game states besides 'game over'
and 'in game'. Menus are technically included, but no UI work for this
has been done.

This commit should also resolve a bug where the UI wouldn't be displayed
on a real device because the draw call returned false, despite it
appearing OK on the simulator.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>