SPRPD-6: Draw UI dialogs last
Due to a render priority issue, UI dialogs wouldn't consistently be
displayed, appearing as if they never rendered at all unless the
menu was opened on device.

To fix this, a new drawUI call was added to GameSystem to allow the
game to render UI at the highest priority, regardless of other
subsystems. This also consistently works in the simulator.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>