SPRPD-7: Match crank and D-Pad movement
The first builds of the game directly tied movement to the crank's
exact angle, which made it feel very fluid and responsive. However,
this gave those that used the crank a very unfair advantage, as they
can warp around the screen edges and move significantly faster than
those that prefered to use the D-Pad.

To fix this, a few changes were made:
- First, the movement function now updates at a fixed point, rather
  than purely from rotation. This allows it to behave the same on both
  the D-Pad and the crank.
- Second, the game now passes over the crank change rather than the
  raw rotation. This allows the movement function to make better
  informed decisions on which way the crank is going, rather than trying
  to guess based on the existing angle.
- Finally, the horizontal position clamps were increased to account
  for the edges of the map. This should effectively make it possible
  to hit the wall rather than loop around.

The tests for the updated function are going to fail right now, but
I can't be bothered to try and get the tests running again since it
is VERY unhappy that PlaydateKit is in the same repo, for some reason.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>